refactor: extracted variant plugins (#47)
